refactor: 스크롤Y값 전달받는 부분을 추상화 (hook함수화)
